Organize corpora by name instead of initial usage
Text corpora can be used for different tasks. The current API mandates the use
for a particular task by coupling the raw dataset representation with the the
transforms. This re-organizes the code declaring the different datasets as a
preparation for refactoring the API.
